LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-badInput-expected.txt


PASS [INPUT in EMAIL status] The multiple attribute is false and the value attribute is empty
PASS [INPUT in EMAIL status] The multiple attribute is false and the value attribute is a valid e-mail address
PASS [INPUT in EMAIL status] The multiple attribute is true and the value contains valid e-mail addresses
PASS [INPUT in EMAIL status] The multiple attribute is true and the value attribute contains a ','
FAIL [INPUT in DATETIME-LOCAL status] The datetime-local type must be supported. assert_equals: The datetime-local type should be supported. expected "datetime-local" but got "text"
FAIL [INPUT in COLOR status] The color type must be supported. assert_equals: The color type should be supported. expected "color" but got "text"
